change generate_schemas to /schemas api route, avoid committing generated json
